refactor: fix cache to disk

+++ src/timeseriesflattener/feature_specs/single_specs.py
@@ -87,7 +87,7 @@ class PredictorSpec:
     base_values_df: pd.DataFrame
     feature_base_name: str
     aggregation_fn: Callable
-    fallback: Union[str, int]
+    fallback: Union[str, float]
     lookbehind_days: float = LOOKBEHIND_DAYS_DEF
     prefix: str = "pred"
